requirements.txt
aiofiles==24.1.0
fastapi==0.111.0
langchain==0.2.6
PyMuPDF
langchain_community==0.2.6
langchain_core==0.2.11
langchain_openai==0.1.14
langchain_text_splitters==0.2.2
langgraph==0.1.5

openai==1.58.1
python-dotenv==1.0.1
rapidfuzz==3.9.3
Requests==2.32.3
typing_extensions==4.12.2
faiss-cpu
aiosqlite
asyncio